Consolidate grid areas query logic
