+++ apps/dh/api-dh/source/DataHub.WebApi/GraphQL/Query/GridAreaQuery.cs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 using Energinet.DataHub.WebApi.Clients.MarketParticipant.v1;
+using Energinet.DataHub.WebApi.GraphQL.Extensions;
 
 namespace Energinet.DataHub.WebApi.GraphQL.Query;
 
@@ -23,29 +24,6 @@ public partial class Query
         await client.GridAreaOverviewAsync();
 
     public async Task<IEnumerable<GridAreaDto>> GetGridAreasAsync(
-        [Service] IMarketParticipantClient_V1 client)
-    {
-        var actors = await client.ActorGetAsync();
-        var gridAreas = await client.GridAreaGetAsync();
-        return gridAreas.Select(gridArea =>
-        {
-            var owner = actors.FirstOrDefault(actor =>
-                actor.Status == "Active" &&
-                actor.MarketRoles.Any(mr =>
-                    mr.EicFunction == EicFunction.GridAccessProvider &&
-                    mr.GridAreas.Any(ga => ga.Id == gridArea.Id)));
-
-            return owner == null
-                ? gridArea
-                : new GridAreaDto
-                {
-                    Id = gridArea.Id,
-                    Code = gridArea.Code,
-                    Name = owner.Name.Value,
-                    PriceAreaCode = gridArea.PriceAreaCode,
-                    ValidFrom = gridArea.ValidFrom,
-                    ValidTo = gridArea.ValidTo,
-                };
-        });
-    }
+        [Service] IMarketParticipantClient_V1 client) =>
+        await client.GetGridAreasAsync();
 }
